cvi_mpi: weekly rls 2024.10.14
-81c8170, Merge "[fix](cvi_mpi): fix build error" into v4.1.0
-8b37430, [fix](cvi_mpi): fix build error
-0834a4c, Merge "[fix](ut): Remove unsupport case
-e042799, Merge "[refactor]( cvi_comm_aio.h): Invalid interface cleaning & interface adding
-c07cbb4, [fix](cvi_mpi): fix rls_info.txt
-9fea320, Merge "[fix](cvi_mpi):Modify print type" into v4.1.0
-bc2babc, [fix](cvi_mpi):Modify print type
-b6bf737, [fix](cvi_mpi): modify .gitignore
...
-2428d0b, [feat](vpss):add macro definition test & structure attribute test
-d8e85d7, [feat](ive): support sub shift mode

+++ cvi_mpi/modules/ive/src/cvi_ive.c
@@ -193,7 +193,7 @@ void dump_data_u64(const CVI_U64 *data, CVI_S32 count, const CVI_CHAR *desc)
 	CVI_S32 i;
 	printf("%s\n", desc);
 	for (i = 0; i < count; ++i) {
-		printf("%016jX ", data[i]);
+		printf("%016jX ", (uintmax_t)data[i]);
 		if ((i + 1) % 2 == 0 || i + 1 == count) {
 			printf(" ");
 			if ((i + 1) % 4 == 0) {
@@ -662,7 +662,7 @@ CVI_S32 CVI_IVE_CompareIveImage(IVE_IMAGE_S *pstImage1, IVE_IMAGE_S *pstImage2)
 				printf("Compare failed, at y = %d, x = %d, n = %d\n",
 					   y, idx, n);
 				printf("  A = 0x%016jX, B = 0x%016jX\n",
-					   pData1_U64[idx], pData2_U64[idx]);
+					   (uintmax_t)pData1_U64[idx], (uintmax_t)pData2_U64[idx]);
 				dump_data_u64(pData1_U64, u32Width, "A");
 				dump_data_u64(pData2_U64, u32Width, "B");
 				s32Succ = CVI_FAILURE;
